wwwmyforumru/requirements.txt
Django==1.7.1
django-registration==1.0